pyproject.toml
[project]
name = "graduatework"
version = "0.1.0"
description = ""
authors = [
    {name = "Aleksei Prusikov",email = "imemphatic@yandex.ru"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django (>=5.1.6,<6.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "djangorestframework (>=3.15.2,<4.0.0)",
    "djangorestframework-simplejwt (>=5.4.0,<6.0.0)",
    "django-cors-headers (>=4.6.0,<5.0.0)",
    "drf-yasg (>=1.21.8,<2.0.0)",
    "pillow (>=11.1.0,<12.0.0)",
    "tzdata (>=2025.1,<2026.0)",
    "sqlparse (>=0.5.3,<0.6.0)",
    "pyyaml (>=6.0.2,<7.0.0)",
    "pytz (>=2025.1,<2026.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "coverage (>=7.6.10,<8.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
isort = "^6.0.0"
black = "^25.1.0"
pytest = "^8.3.4"
pytest-django = "^4.9.0"
pytest-cov = "^6.0.0"
flake8 = "^7.1.1"

